goliath_daily/requirements.txt
python-dotenv>=1.0.0
PyMySQL>=1.1.0
